GL4: Multithreaded Renderer
This multithreads certain parts of the renderer.  The parts dealing with images and particles in particular, as these generate a lot of CPU overhead, especially on less powerful machines.  I also have severely enhanced the efficiency of the renderer itself, resulting in over 100% performance gains depending on hardware.

libpthread is required on all UNIX-like systems.  On Windows, the threading library is part of -lkernel32